Import the script from the TurboGears docs
The script was taken at the following page:
    http://turbogears.org/1.0/docs/Identity/Recipes.html#authenticate-against-an-ldap-server-updated

Unfortunately, there is no licensing information. The TurboGears
community usually release their stuff under the MIT license (as
is the case of TurboGears itself), which is also the license I've
chosen for tgldapid.

Hopefully, this won't cause any problem. If you are the original
author, please let me know whether you're ok with this. At the
very least, I would like to give you credit for your work.