Stop calling the parent validate_identity method
This has several benefits.

First, if we call validate_password ourselves, then we can pass it
the LDAP record for the user.

That, in turn, means that the validate_password method doesn't need
to do one more connection to the LDAP server to get informations we
already obtained.

All in all, it just simplifies the code.